implementando respostas adequadas para os endpoints

+++ src/main/java/com/datapro/apiSinopsers/service/RealityShowService.java
@@ -4,6 +4,7 @@ import com.datapro.apiSinopsers.dto.RealityShowDto;
 import com.datapro.apiSinopsers.model.RealityShow;
 import com.datapro.apiSinopsers.repository.RealityShowRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -14,11 +15,18 @@ public class RealityShowService {
     @Autowired
     private RealityShowRepository realityShowRepository;
 
-    public void cadastrarReality(RealityShowDto realityShowDto) {
-        RealityShow realityShow = new RealityShow();
-        realityShow.setTitulo(realityShowDto.getTitulo());
-        realityShow.setSinopse(realityShowDto.getSinopse());
-        realityShowRepository.save(realityShow);
+    private static final String REALITY_NAO_ENCONTRADO = "Realityshow não encontrado!";
+
+    public ResponseEntity cadastrarReality(RealityShowDto realityShowDto) {
+        String titulo =  realityShowRepository.buscarPorTitulo(realityShowDto.getTitulo());
+        if (titulo == null){
+            RealityShow realityShow = new RealityShow();
+            realityShow.setTitulo(realityShowDto.getTitulo());
+            realityShow.setSinopse(realityShowDto.getSinopse());
+            realityShowRepository.save(realityShow);
+            return ResponseEntity.status(201).body("Realityshow cadastrado com sucesso!");
+        }
+       return ResponseEntity.status(409).body("Realityshow ja cadastrado seja mais criativo!");
     }
 
 
@@ -36,15 +44,27 @@ public class RealityShowService {
         return listaRealitys;
     }
 
-    public void atualizarReality(RealityShowDto realityShowDto){
+    public ResponseEntity atualizarReality(RealityShowDto realityShowDto){
+
         RealityShow realityShow = realityShowRepository.buscarPorId(realityShowDto.getId());
-        realityShow.setTitulo(realityShowDto.getTitulo());
-        realityShow.setSinopse(realityShowDto.getSinopse());
-        realityShowRepository.save(realityShow);
+        if (realityShow != null) {
+            realityShow.setTitulo(realityShowDto.getTitulo());
+            realityShow.setSinopse(realityShowDto.getSinopse());
+            realityShowRepository.save(realityShow);
+            return ResponseEntity.status(200).body("Realityshow atualizado com sucesso!");
+        }
+        return ResponseEntity.status(409).body(REALITY_NAO_ENCONTRADO);
     }
 
-    public void excluirReality(Long id){
-        realityShowRepository.deleteById(id);
+    public ResponseEntity excluirReality(Long id){
+        RealityShow realityShow = realityShowRepository.buscarPorId(id);
+        if (realityShow != null){
+            realityShowRepository.deleteById(id);
+            return ResponseEntity.status(204).body("Realityshow deletado com sucesso!");
+        }
+
+        return ResponseEntity.status(409).body(REALITY_NAO_ENCONTRADO);
     }
+
 }
